+++ src/io/io.h
@@ -15,8 +15,8 @@ namespace arboretum {
       std::vector<float> y_hat;
       std::vector<float> y;
       std::vector<float> grad;
-      int rows;
-      int columns;
+      size_t rows;
+      size_t columns;
       void Init(const float initial_y, std::function<float const(const float, const float)> func);
       void UpdateGrad();
       DataMatrix(int rows, int columns);
